Разделил объявления и описания по разным файлам.
